pyproject.toml
[project]
name = "biases-in-the-blind-spot"
version = "0.1.0"
description = "Anonymized code for the paper 'Biases in the Blind Spot: Detecting What LLMs Fail to Mention'"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "chat-limiter>=0.15.0",
    "dataclass-wizard>=0.35.1",
    "dataclasses-json>=0.6.7",
    "editables>=0.5",
    "hatchling>=1.27.0",
    "hf-transfer>=0.1.9",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "kagglehub>=0.3.13",
    "names-generator>=0.2.0",
    "nbformat>=5.10.4",
    "openai>=1.90.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "redlines>=0.5.2",
    "scikit-learn>=1.0.0",
    "scipy>=1.10.0",
    "statsmodels>=0.14.0",
    "torch",
    "torchcodec>=0.6.0",
    "transformers>=4.0.0",
    "tueplots>=0.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
exclude = [".venv", "experiments"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501", "E731", "E722", "F722", "F821"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["biases_in_the_blind_spot/**/*.py", "tests/**/*.py"]
exclude = ["**/__pycache__", ".venv", "experiments"]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingImports = "warning"
reportMissingTypeStubs = false

[dependency-groups]
dev = ["pyright>=1.1.402", "pytest>=8.4.1", "ruff>=0.12.0"]
